fixing english last name bug
